Create and consume roles to tighten screws
The whole module was creating and using objects in manners that did not
match the classes.

It meant that creating an alternative link provider was unreliable,
because there was no documentation on any of it, nor any code-based
constraints.

This creates roles for the classes to consume, and that we can consume
elsewhere to improve compatibility.