Remove key fields for relationships in create even if undef
Odoo 14 is a lot stricter on create() and will throw an exception if you
pass a field that isn't defined on the model in question.
So if the relationship name isn't the same as the key field on the Odoo
side, we must remove the key field from the hash we send to Odoo.